runtime.Run without file
